refactor(order-import): sửa một chút feedback

+++ src/main/java/com/rk/WMS/order/criteria/SearchOrderCriteria.java
@@ -13,5 +13,5 @@ public class SearchOrderCriteria {
   String supplierPhone;
   String receiverPhone;
   Integer statusCode;
-  Integer warehouseId;
+  Long warehouseId;
 }
